+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
-See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+All notable changes to this project will be documented in this file. See
+[standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+### 0.2.0 (2026-07-17)
+
+#### Add
+
+- Add new method `getAddressOperations`
+
+#### Hotfix
+
+- Properly parse the referral data before sending it to the API
+- Do not force users to provide an object when creating the instance of the kit
 
 ### 0.1.0 (2026-06-10)